Move SamlSerializer to correct namespace
